Move all passes into a separate module. (#638)
